+++ explorer/node/db.js
@@ -4,6 +4,8 @@ const { ThemeIcon, ThemeColor, TreeItemCollapsibleState } = require("vscode")
 const { RedisModel } = require("../../command/redis")
 const { RedisDateTypes } = require("./type")
 const { log } = require("../../lib/logging")
+const { KeyTreeItem } = require("./key")
+const { isEmpty } = require("../../lib/util")
 
 
 class ShowMoreKeysTreeItem extends TreeDataItem {
@@ -26,6 +28,36 @@ class ShowMoreKeysTreeItem extends TreeDataItem {
     return new ShowMoreKeysTreeItem(opt)
   }
 }
+
+class SearchKeysTreeItem extends TreeDataItem {
+  constructor(opt = {}) {
+    super(opt)
+    // this.isShowMoreItem = true
+
+
+  }
+
+  static init(opt = {}) {
+    opt.contextValue = NodeType.SEARCHKEY
+    opt.label = 'search key'
+    opt.tooltip = `Search result`
+    opt.collapsibleState = TreeItemCollapsibleState.Expanded
+    return new SearchKeysTreeItem(opt)
+  }
+  async getChildren() {
+    const { db, connection, redisModel, } = this
+    const redisDataType = 'searchResult'
+    const item = this.redisModel.searchResult
+    return this.redisModel.searchResult.map(({ key: label, type }) => {
+      return KeyTreeItem.init({
+        label, redisDataType,
+        db, connection, redisModel,
+        tooltip: `${type}`
+      })
+    })
+  }
+}
+
 class DbTreeItem extends TreeDataItem {
   constructor(opt = {}) {
     super(opt)
@@ -58,6 +90,12 @@ class DbTreeItem extends TreeDataItem {
         collapsibleState: TreeItemCollapsibleState.Collapsed
       })
     })
+    let searchItem = null
+    if (this.redisModel && !isEmpty(this.redisModel.searchResult)) {
+      searchItem = SearchKeysTreeItem.init({
+        db, connection, redisModel
+      })
+    }
     let scanMoreItem = null
     if (scanMore !== +keysLen) {
       scanMoreItem = ShowMoreKeysTreeItem.init({
@@ -66,7 +104,7 @@ class DbTreeItem extends TreeDataItem {
         refreshParent: this
       })
     }
-    return [...categroys, scanMoreItem]
+    return [...categroys, searchItem, scanMoreItem]
   }
 }